chore: remove unused circuit append with subtrees
