+++ prover/client/src/lib.rs
@@ -1,6 +1,5 @@
 pub mod batch_address_append;
 pub mod batch_append_with_proofs;
-pub mod batch_append_with_subtrees;
 pub mod batch_update;
 pub mod combined;
 pub mod combined_legacy;
