+++ prover/server/prover/circuit_utils.go
@@ -220,3 +220,37 @@ func (gadget MerkleRootUpdateGadget) DefineGadget(api frontend.API) interface{} 
 	})
 	return newRoot
 }
+
+// incrementBits implements binary addition to increment a number represented as a list of bits.
+// It uses XOR and AND operations to efficiently increment the binary number without using
+// traditional arithmetic operations, which is beneficial in zero-knowledge proof circuits.
+//
+// The function works as follows:
+//  1. It starts with a carry of 1 (equivalent to adding 1 to the number).
+//  2. For each bit, from least to most significant:
+//     a. It XORs the current bit with the carry. This effectively adds the bit and carry
+//     without considering a new carry. (0⊕0=0, 0⊕1=1, 1⊕0=1, 1⊕1=0)
+//     b. It ANDs the original bit with the carry to determine if there should be a carry
+//     for the next bit. (0∧0=0, 0∧1=0, 1∧0=0, 1∧1=1)
+//     c. The result of XOR becomes the new value for the current bit.
+//     d. The result of AND becomes the new carry for the next iteration.
+//  3. This process continues for all bits, resulting in the incremented binary number.
+//
+// Example: Incrementing 0111 (7 in decimal)
+// Initial state: 0111, carry = 1
+// i=0: 1⊕1=0, carry=1∧1=1 -> 0110, carry=1
+// i=1: 1⊕1=0, carry=1∧1=1 -> 0010, carry=1
+// i=2: 1⊕1=0, carry=1∧1=1 -> 1010, carry=1
+// i=3: 0⊕1=1, carry=0∧1=0 -> 1000, carry=0
+// Final result: 1000 (8 in decimal)
+func incrementBits(api frontend.API, bits []frontend.Variable) []frontend.Variable {
+	carry := frontend.Variable(1)
+	for i := 0; i < len(bits); i++ {
+		// XOR operation implements binary addition without carry
+		newBit := api.Xor(bits[i], carry)
+		// AND operation determines if we need to carry to the next bit
+		carry = api.And(bits[i], carry)
+		bits[i] = newBit
+	}
+	return bits
+}
